only copy asset has substring `added` by proxyAdd.php

+++ proxyChecker.php
@@ -123,14 +123,16 @@ if (gethostname() !== 'DESKTOP-JVTSJ6I') {
   $assets = getFilesByExtension(__DIR__ . '/assets/proxies', 'txt');
   $assets[] = __DIR__ . '/proxies-backup.txt';
   foreach ($assets as $asset) {
-    echo $asset . ' - ' . (file_exists($asset) ? 'true' : 'false') . PHP_EOL;
-    if (file_exists($asset)) {
-      $move = moveContent($asset, $untestedFilePath);
-      if ($move === '') {
-        unlink($asset);
-        exit('copied assets');
-      } else {
-        echo $move . PHP_EOL;
+    if (strpos($asset, 'added-') !== false) {
+      echo $asset . ' - ' . (file_exists($asset) ? 'true' : 'false') . PHP_EOL;
+      if (file_exists($asset)) {
+        $move = moveContent($asset, $untestedFilePath);
+        if ($move === '') {
+          unlink($asset);
+          exit('copied assets');
+        } else {
+          echo $move . PHP_EOL;
+        }
       }
     }
   }
